refactor(diagnose): debloat — drop redundant modeling, split by concern
What
- Remove the redundant Severity (warn/info) axis: after Kind (problem/advisory)
  it was a second, ~90%-correlated axis the renderer never printed. Order hints
  by Kind alone.
- Remove dead Decision.When / BlockedHost.LastSeen (computed, never rendered;
  dropped the time dependency from the pure package).
- Collapse profileSummaryV + profileSummary() into the existing diagnose.Policy
  (one profile->struct mapping instead of two near-duplicates).
- Split cli/diagnose.go (592 LOC) into diagnose.go (retrospective analysis +
  render, 357) and diagnose_probe.go (all probe variants, 254) so the two
  concerns review independently.
- Make --json consistent: add snake_case tags to diagnose.Report/Policy/Hint/
  BlockedHost (previously untagged → mixed-case output; tests passed only via
  Go's case-insensitive unmarshal).

Why
- Self-review round: cut redundant abstraction and dead code, right-size files
  for a human reviewer, and fix a half-tagged public JSON contract.

Verification
- go test ./... green; gofmt/vet clean; behavior unchanged (focused view,
  --json now uniform snake_case). Tests updated to assert on Kind, not Severity.
